Move ConfigTest to test_objects.py.
